terraform/streamlit/requirements.txt
streamlit
google-cloud-bigquery
google-cloud-pubsub
pandas
pydeck
streamlit-autorefresh 